requirement.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cloudinary==1.24.0
cryptography==3.3.1
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==3.1.4
django-cloudinary-storage==0.3.0
django-crispy-forms==1.10.0
gunicorn==20.0.4
idna==2.10
mysqlclient==2.0.2
oauthlib==3.1.0
Pillow==8.0.1
pkg-resources==0.0.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.0
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
